pyproject.toml
[tool.poetry]
name = "reusable-fastapi"
version = "0.1.0"
description = ""
authors = ["dev-bong <qhddnjswo011@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
black = "^24.4.2"
pydantic = {extras = ["email"], version = "^2.7.4"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
